se ha agregado la lógica para las axn a implementar en las axn del parser, se procederá a addlas y a realizar pruebas, además de add la division para mostrar #linea y #columna en la interfaz

+++ GraficadorEstadistico/app/src/main/java/Backend/Objetos/Graficas/Barras.kt
@@ -1,9 +1,8 @@
 package Backend.Objetos.Graficas
 
-class Barras(titulo: String, unir: Array<IntArray>, ejeX:Array<String>, ejeY:IntArray) : Grafica(titulo, unir) {
+class Barras(titulo: String, unir: Array<IntArray>, ejeX:Array<String>, ejeY:DoubleArray) : Grafica(titulo, unir) {
     var ejeX:Array<String> = ejeX//los dejé como array pensando en que son equivalentes a los arreglos en java [ya que convertiré las listas creadas en el parser a un array...
-    var ejeY:IntArray = ejeY
-
+    var ejeY:DoubleArray = ejeY
 
     //cree los getter pensando en el hecho de que son variables private, pero aquí los getter y setter son como atribs de las variables, entonces si la var es privada, estos tb lo serán, por eso es que Kotlin los crea por defecto, porque a ellos se accede a travez de la var/atrib...
 }
